att seguranca - att metodo de pagamento funcionando

+++ backend/app/models/servico.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.orm import synonym
 from app.database import Base
 
@@ -11,10 +11,10 @@ class Servico(Base):
     id = Column(Integer, primary_key=True, index=True)
     nome = Column(String(255), nullable=False)
     duracao_minutos = Column(Integer, nullable=False)
-    preco = Column(Float, nullable=False)
+    preco = Column(Numeric(12, 2), nullable=False)
     pagamento_adiantado_obrigatorio = Column(Boolean, nullable=False, default=False)
     advance_payment_type = Column(String(20), nullable=True)
-    advance_payment_amount = Column(Float, nullable=True)
+    advance_payment_amount = Column(Numeric(12, 2), nullable=True)
     payment_description_override = Column(Text, nullable=True)
     updated_at = Column(DateTime, nullable=False, default=datetime.utcnow, onupdate=datetime.utcnow)
     estabelecimento_id = Column(Integer, ForeignKey("estabelecimentos.id"), nullable=True, index=True)
